refactor : view로 패키지 이동
