+++ src/test/java/chess/domain/CommandTest.java
@@ -1,6 +1,6 @@
 package chess.domain;
 
-import chess.domain.Command;
+import chess.view.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
